+++ appseed/math/kissfft/kiss_fft.c
@@ -239,7 +239,7 @@ static void kf_bfly_generic(
             int twidx=0;
             Fout[ k ] = scratchbuf[0];
             for (q=1;q<p;++q ) {
-                twidx += fstride * k;
+                twidx += (int) (fstride * k);
                 if (twidx>=Norig) twidx-=Norig;
                 C_MUL(t,scratchbuf[q] , twiddles[twidx] );
                 C_ADDTO( Fout[ k ] ,t);
